requirements.txt
annotated-types==0.7.0
astroid==4.0.4
basedpyright==1.39.0
beautifulsoup4==4.13.4
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
dill==0.4.1
Flask==3.1.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
greenlet==3.3.2
gunicorn==23.0.0
idna==3.11
isort==8.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
mailtrap==2.5.0
MarkupSafe==3.0.3
mccabe==0.7.0
nodejs-wheel-binaries==24.14.1
packaging==26.0
pillow==12.2.0
platformdirs==4.9.4
pydantic==2.12.5
pydantic_core==2.41.5
pylint==4.0.5
python-dotenv==1.2.2
requests==2.33.0
soupsieve==2.8.3
SQLAlchemy==2.0.48
tomlkit==0.14.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.7
